# perf(gatsby): drop severe scaling regression caused by analytics