@pvdz 2020-06-02
While this is already a good fix, I think I need to dig a little deeper. According to old benchmark dumps this is what master on February 12th did for this benchmark:

```
info bootstrap finished - 115.315 s
success Building production JavaScript and CSS bundles - 5.162s
success run queries - 55.484s - 145178/145178 2616.58/s
success Building static HTML for pages - 113.598s - 145178/145178 1278.00/s
info Done building in 292.68 sec
Done in 298.79s.
```

So it seems we could still pump 50% faster. Hopefully that regression is also fairly easy to find.