@pvdz 2020-06-02
Went to repo tag `gatsby@2.19.17`:
```
info bootstrap finished - 86.401 s                                                                                                                                                
success Building production JavaScript and CSS bundles - 8.012s                                                                                                                   
success run queries - 78.374s - 150002/150002 1913.92/s                                                                                                                           
success Building static HTML for pages - 140.455s - 150002/150002 1067.97/s                                                                                                       
info Done building in 312.881 sec                                                                                                                                                 
```

And repo at tag `gatsby@2.19.16`:

```
info bootstrap finished - 88.241 s                                                                                                                                                
success Building production JavaScript and CSS bundles - 6.144s                                                                                                                   
success run queries - 75.819s - 150002/150002 1978.43/s                                                                                                                           
success Building static HTML for pages - 140.238s - 150002/150002 1069.63/s                                                                                                       
info Done building in 312.954 sec                       
```

Hard to tell what the regression might have been or how I was able to get the perf mentioned above. Could be a plugin that wasn't bumped or ... I'm not sure. But I'm pretty sure it's not worth chasing this kinda of delta over such a long timespan (over 3 months) so let's look forward.